requirements.txt
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-vision==2.2.0
googleapis-common-protos==1.53.0
python-dotenv==0.15.0